+++ app/src/main/java/com/effeta/miparroquiaandroid/repositories/AnnouncementRepository.kt
@@ -1,10 +1,8 @@
 package com.effeta.miparroquiaandroid.repositories
 
-import android.arch.lifecycle.Observer
-import android.util.Log
 import com.effeta.miparroquiaandroid.models.Announcement
 import com.effeta.miparroquiaandroid.services.firebase.FirebaseAnnouncement
-import io.reactivex.*
+import io.reactivex.Observable
 
 /**
  * Created by aulate on 16/2/18.
@@ -12,22 +10,6 @@ import io.reactivex.*
 object AnnouncementRepository {
 
     fun getAnnouncements() : Observable<List<Announcement>> {
-        return Observable.create<List<Announcement>> {
-            FirebaseAnnouncement.getAnnouncementList().addOnCompleteListener { task ->
-                if (task.isSuccessful) {
-                    val list = ArrayList<Announcement>()
-                    task.result.documents.forEach({ documentSnapshot ->
-                        val a = documentSnapshot.toObject(Announcement::class.java)
-                        a.mKey = documentSnapshot.id
-                        Log.d("AnnouncementRepository", a.toString())
-                        list.plus(a)
-                    })
-                    it.onNext(list)
-                    it.onComplete()
-                } else {
-                    it.onError(task.exception!!)
-                }
-            }
-        }
+        return FirebaseAnnouncement.getAnnouncementList("1")
     }
 }
