+++ app/src/main/java/com/effeta/miparroquiaandroid/viewmodel/AnnouncementViewModel.kt
@@ -1,6 +1,5 @@
 package com.effeta.miparroquiaandroid.viewmodel
 
-import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
 import com.effeta.miparroquiaandroid.models.Announcement
